# api/query/openapi.yaml
openapi: 3.0.3
info:
  title: Timadorus Query API
  version: 0.1.0
  description: >
    Read-side (query) API for the Timadorus CQRS/ES platform. Every operation requires a
    bearer token. The remaining hierarchy lands in later phases (see the implementation
    plan).
servers:
  - url: /

security:
  - bearerAuth: []

paths:
  /users/{userId}:
    get:
      operationId: getUser
      summary: Get a User by id.
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: The User.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"

  /universes/{universeId}:
    get:
      operationId: getUniverse
      summary: Get a Universe by id.
      parameters:
        - $ref: "#/components/parameters/UniverseId"
      responses:
        "200":
          description: The Universe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Universe"
        "404":
          $ref: "#/components/responses/NotFound"

  /universes/{universeId}/creators:
    get:
      operationId: listUniverseCreators
      summary: List a Universe's Creators.
      parameters:
        - $ref: "#/components/parameters/UniverseId"
      responses:
        "200":
          description: Creator user ids.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  format: uuid
        "404":
          $ref: "#/components/responses/NotFound"

  /universes/{universeId}/campaigns:
    get:
      operationId: listCampaignsByUniverse
      summary: List non-archived Campaigns under a Universe.
      parameters:
        - $ref: "#/components/parameters/UniverseId"
      responses:
        "200":
          description: Campaigns.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"

  /universes/{universeId}/entities:
    get:
      operationId: listEntitiesByUniverse
      summary: List non-archived Entities under a Universe, optionally filtered by name.
      parameters:
        - $ref: "#/components/parameters/UniverseId"
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: >-
            Case-insensitive substring match on name. When present, results are capped at 20
            and ordered by name; when absent, behavior is unchanged (full non-archived list).
      responses:
        "200":
          description: Entities.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entity"

  /universes/{universeId}/objects:
    get:
      operationId: listObjectsByUniverse
      summary: List non-archived Objects under a Universe, optionally filtered by name.
      parameters:
        - $ref: "#/components/parameters/UniverseId"
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: >-
            Case-insensitive substring match on name. When present, results are capped at 20
            and ordered by name; when absent, behavior is unchanged (full non-archived list).
      responses:
        "200":
          description: Objects.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Object"

  /universes/{universeId}/changes/cursor:
    get:
      operationId: getUniverseChangesCursor
      summary: Get the current change cursor for a Universe.
      parameters:
        - $ref: "#/components/parameters/UniverseId"
      responses:
        "200":
          description: The current cursor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UniverseChangeCursor"

  /universes/{universeId}/changes:
    get:
      operationId: listUniverseChanges
      summary: List changes to this Universe and everything inside it, after a given cursor.
      parameters:
        - $ref: "#/components/parameters/UniverseId"
        - name: since
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Changes, ordered by globalSeq ascending, capped at 20 per call.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UniverseChange"

  /campaigns/{campaignId}:
    get:
      operationId: getCampaign
      summary: Get a Campaign by id.
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "200":
          description: The Campaign.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          $ref: "#/components/responses/NotFound"

  /campaigns/{campaignId}/gamemasters:
    get:
      operationId: listCampaignGamemasters
      summary: List a Campaign's Gamemasters.
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "200":
          description: Gamemaster user ids.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  format: uuid
        "404":
          $ref: "#/components/responses/NotFound"

  /campaigns/{campaignId}/characters:
    get:
      operationId: listCharactersByCampaign
      summary: List non-archived Characters under a Campaign.
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "200":
          description: Characters.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Character"

  /entities/{entityId}:
    get:
      operationId: getEntity
      summary: Get an Entity by id.
      parameters:
        - $ref: "#/components/parameters/EntityId"
      responses:
        "200":
          description: The Entity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
        "404":
          $ref: "#/components/responses/NotFound"

  /characters/{characterId}:
    get:
      operationId: getCharacter
      summary: Get a Character by id.
      parameters:
        - $ref: "#/components/parameters/CharacterId"
      responses:
        "200":
          description: The Character.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "404":
          $ref: "#/components/responses/NotFound"

  /objects/{objectId}:
    get:
      operationId: getObject
      summary: Get an Object by id.
      parameters:
        - $ref: "#/components/parameters/ObjectId"
      responses:
        "200":
          description: The Object.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Object"
        "404":
          $ref: "#/components/responses/NotFound"

  /rulesets:
    get:
      operationId: listRulesets
      summary: List non-archived Rulesets.
      responses:
        "200":
          description: Rulesets.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ruleset"

  /rulesets/{rulesetId}:
    get:
      operationId: getRuleset
      summary: Get a Ruleset by id.
      parameters:
        - $ref: "#/components/parameters/RulesetId"
      responses:
        "200":
          description: The Ruleset.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ruleset"
        "404":
          $ref: "#/components/responses/NotFound"

  /rulesets/{rulesetId}/tables/{tableName}:
    get:
      operationId: listRulesetTableRows
      summary: List every row of a Ruleset's named data table.
      parameters:
        - $ref: "#/components/parameters/RulesetId"
        - name: tableName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Table rows.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RulesetTableRow"

  /rulesets/{rulesetId}/tables/{tableName}/{rowKey}:
    get:
      operationId: getRulesetTableRow
      summary: Get one row's data from a Ruleset's named data table.
      parameters:
        - $ref: "#/components/parameters/RulesetId"
        - name: tableName
          in: path
          required: true
          schema:
            type: string
        - name: rowKey
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The row's opaque, table-specific data.
          content:
            application/json:
              schema:
                type: object
        "404":
          $ref: "#/components/responses/NotFound"

  /users:
    get:
      operationId: listUsers
      summary: List non-archived Users.
      responses:
        "200":
          description: Users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /universes:
    get:
      operationId: listUniverses
      summary: List non-archived Universes.
      responses:
        "200":
          description: Universes.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Universe"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    UniverseId:
      name: universeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CampaignId:
      name: campaignId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    EntityId:
      name: entityId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CharacterId:
      name: characterId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ObjectId:
      name: objectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RulesetId:
      name: rulesetId
      in: path
      required: true
      schema:
        type: string
        format: uuid

  schemas:
    Object:
      type: object
      required: [id, name, universeId, isArchived]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        universeId:
          type: string
          format: uuid
        isArchived:
          type: boolean

    Character:
      type: object
      required: [id, name, campaignId, entityId, playerUserId, info, isArchived]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        campaignId:
          type: string
          format: uuid
        entityId:
          type: string
          format: uuid
        playerUserId:
          type: string
          format: uuid
        # info is an opaque JSON-string payload, never parsed or validated by the backend.
        # Settable only via PUT /characters/{characterId}/info (api/command/openapi.yaml) —
        # not at creation, and not by any other command.
        info:
          type: string
        isArchived:
          type: boolean

    Entity:
      type: object
      required: [id, name, universeId, isArchived]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        universeId:
          type: string
          format: uuid
        isArchived:
          type: boolean

    Campaign:
      type: object
      required: [id, name, universeId, rulesetId, configuration, isArchived]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        universeId:
          type: string
          format: uuid
        rulesetId:
          type: string
          format: uuid
        # configuration is an opaque JSON-string payload, never parsed or validated by the
        # backend. Settable directly via PUT /campaigns/{campaignId}/configuration, or
        # indirectly (as a {"configs":[...]} timestamp list) by timadorus-engine reacting to
        # PUT /campaigns/{campaignId}/configure — see api/command/openapi.yaml. Also populated
        # automatically (as a {"traits":[...]} list, merged in alongside any "configs" list) by
        # timadorus-engine immediately on Campaign creation, with no client action, for Campaigns
        # that use the "timadorus" Ruleset.
        configuration:
          type: string
        isArchived:
          type: boolean

    Ruleset:
      type: object
      required: [id, name, description, references, isArchived]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        references:
          type: array
          items:
            type: string
        isArchived:
          type: boolean

    RulesetTableRow:
      type: object
      required: [key, data]
      properties:
        key:
          type: string
        # data is an opaque, table-specific payload — its shape varies per table (see
        # internal/engine/timadorus/tables for what each table's own Go struct actually contains;
        # there is no single fixed schema across tables, by design — traits.yaml's rows look
        # nothing like a hypothetical future skills table's rows).
        data:
          type: object

    UniverseChangeCursor:
      type: object
      required: [globalSeq]
      properties:
        globalSeq:
          type: integer
          format: int64

    UniverseChange:
      type: object
      required: [globalSeq, aggregateType, aggregateId, eventType, occurredAt]
      properties:
        globalSeq:
          type: integer
          format: int64
        aggregateType:
          type: string
        aggregateId:
          type: string
          format: uuid
        eventType:
          type: string
        occurredAt:
          type: string
          format: date-time

    User:
      type: object
      required: [id, name, isArchived]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        isArchived:
          type: boolean

    Universe:
      type: object
      required: [id, name, isArchived]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        isArchived:
          type: boolean

    Problem:
      description: RFC 7807 problem+json error body.
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        status:
          type: integer

  responses:
    NotFound:
      description: Not found.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"